Added InitSDL and window class. Also changed getters to use const methods
