+++ src/InitSDL.cpp
@@ -4,6 +4,7 @@
 
 #include "InitSDL.h"
 #include <stdlib.h>  
+#include <SDL2/SDL.h>
 
 InitSDL::InitSDL()
 {
@@ -21,6 +22,11 @@ InitSDL::InitSDL()
     glEnable(GL_DEPTH_TEST);
     glEnable(GL_CULL_FACE);
     glCullFace(GL_BACK);
+
+}
+
+InitSDL::~InitSDL()
+{
 
 }
