+++ src/InitSDL.h
@@ -3,7 +3,6 @@
 
 #include <string>
 
-#include <SDL2/SDL.h>
 #include "ModelData.h"
 #include <string>
