+++ src/ObjectLoader.cpp
@@ -8,6 +8,17 @@
 #include "ObjectLoader.h"
 #include "Mesh.h"
 
+
+ObjectLoader::ObjectLoader()
+{
+    data_ = new ModelData();
+}
+
+const ModelData* ObjectLoader::getModel()
+{
+    return data_;
+}
+
 /**
  * Here is a short list of features a real function would provide :
  *  - Binary files. Reading a model should be just a few memcpy's away, not parsing a file at runtime. In short : OBJ is not very great.
@@ -17,8 +28,7 @@
  *  - More secure. Change another line and you can inject code.
  *  - Loading from memory, stream, etc
  */
-
-bool loadObject(const char* path, ModelData& data)
+bool ObjectLoader::loadObject(const char* path)
 {
     std::cout << "Loading OBJ file " << path << "..." << std::endl;
 
@@ -80,22 +90,35 @@ bool loadObject(const char* path, ModelData& data)
             commentLines++;
         }
     }
-    std::vector <glm::vec3>* out_vertices = new std::vector <glm::vec3> ();
-    std::vector <glm::vec2>* out_texCoords = new std::vector <glm::vec2> ();
-    std::vector <glm::vec3>* out_normals = new std::vector <glm::vec3> ();
+	std::vector <glm::vec3>* out_vertices = new std::vector <glm::vec3> ();
+	std::vector <glm::vec2>* out_texCoords = new std::vector <glm::vec2> ();
+	std::vector <glm::vec3>* out_normals = new std::vector <glm::vec3> ();
     for ( unsigned int i = 0; i < vertexIndices.size(); i++ ) // For each vertex get indices of attributes
     {
         unsigned int vertexIndex = vertexIndices[i];
         unsigned int normalIndex = normalIndices[i];
 
         out_vertices->push_back(temp_vertices[vertexIndex - 1]); // Get the attributes thanks to the index
-        out_normals->push_back(temp_normals[normalIndex - 1]);
-        out_texCoords->push_back(glm::vec2(0.2, 0.2)); //Set the texture coordinate to an arbitrary value
+		out_normals->push_back(temp_normals[normalIndex - 1]);
+		out_texCoords->push_back(glm::vec2(0.2, 0.2)); //Set the texture coordinate to an arbitrary value
     }
-    data.setPos(out_vertices);// Put the attributes in buffers
-    data.setNormal(out_normals);
-    data.setTexCoord(out_texCoords);
-    data.setSize(data.getPos().size());
+	data_->setPos(out_vertices);// Put the attributes in buffers
+	data_->setNormal(out_normals);
+	data_->setTexCoord(out_texCoords);
+	data_->setSize(data_->getPos().size());
+
+    std::vector<unsigned int>* indices = new std::vector<unsigned int>();  
+    for (unsigned int i = 0; i < data_->getSize(); i++)
+    {
+        indices->push_back(i);
+    }
+    data_->setIndices(indices);
+
     std::cout << "The .obj file was loaded successfully!" << std::endl;
     return true;
+}
+
+ObjectLoader::~ObjectLoader()
+{
+
 }
